# kura/explorer/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kura Explorer API",
    "description": "API for exploring Kura checkpoint data",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/insights/frustration-map": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Get Frustration Heatmap",
        "description": "Get user frustration heatmap data for clusters.",
        "operationId": "get_frustration_heatmap_api_insights_frustration_map_get",
        "parameters": [
          {
            "name": "min_conversations",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Minimum conversations in cluster",
              "default": 5,
              "title": "Min Conversations"
            },
            "description": "Minimum conversations in cluster"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FrustrationMapItem"
                  },
                  "title": "Response Get Frustration Heatmap Api Insights Frustration Map Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/insights/metadata-dist": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Get Metadata Distribution",
        "description": "Get distribution of metadata values.",
        "operationId": "get_metadata_distribution_api_insights_metadata_dist_get",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Metadata key to analyze",
              "title": "Key"
            },
            "description": "Metadata key to analyze"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetadataDistributionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/insights/themes": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Get Common Themes",
        "description": "Extract common themes across clusters.",
        "operationId": "get_common_themes_api_insights_themes_get",
        "parameters": [
          {
            "name": "min_frequency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Minimum frequency for theme",
              "default": 3,
              "title": "Min Frequency"
            },
            "description": "Minimum frequency for theme"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThemeInsight"
                  },
                  "title": "Response Get Common Themes Api Insights Themes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/insights/outliers": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Get Outlier Conversations",
        "description": "Find conversations that don't fit typical patterns.",
        "operationId": "get_outlier_conversations_api_insights_outliers_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlierResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/insights/compare-clusters": {
      "post": {
        "tags": [
          "insights"
        ],
        "summary": "Compare Clusters",
        "description": "Compare multiple clusters side by side.",
        "operationId": "compare_clusters_api_insights_compare_clusters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "title": "Cluster Ids"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompareResponse": {
        "properties": {
          "clusters": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Clusters"
          }
        },
        "type": "object",
        "required": [
          "clusters"
        ],
        "title": "CompareResponse",
        "description": "Response for cluster comparison."
      },
...
      "FrustrationMapItem": {
        "properties": {
          "cluster_id": {
            "type": "string",
            "title": "Cluster Id"
          },
          "cluster_name": {
            "type": "string",
            "title": "Cluster Name"
          },
          "avg_frustration": {
            "type": "number",
            "title": "Avg Frustration"
          },
          "conversation_count": {
            "type": "integer",
            "title": "Conversation Count"
          },
          "x_coord": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X Coord"
          },
          "y_coord": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y Coord"
          }
        },
        "type": "object",
        "required": [
          "cluster_id",
          "cluster_name",
          "avg_frustration",
          "conversation_count"
        ],
        "title": "FrustrationMapItem",
        "description": "Frustration score mapping."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetadataDistributionResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "total_conversations": {
            "type": "integer",
            "title": "Total Conversations"
          },
          "conversations_with_key": {
            "type": "integer",
            "title": "Conversations With Key"
          },
          "value_distribution": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Value Distribution"
          },
          "unique_values": {
            "type": "integer",
            "title": "Unique Values"
          }
        },
        "type": "object",
        "required": [
          "key",
          "total_conversations",
          "conversations_with_key",
          "value_distribution",
          "unique_values"
        ],
        "title": "MetadataDistributionResponse",
        "description": "Metadata distribution analysis response."
      },
      "OutlierResponse": {
        "properties": {
          "high_frustration": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "High Frustration"
          },
          "very_short": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Very Short"
          },
          "multilingual": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Multilingual"
          },
          "assistant_errors": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Assistant Errors"
          }
        },
        "type": "object",
        "required": [
          "high_frustration",
          "very_short",
          "multilingual",
          "assistant_errors"
        ],
        "title": "OutlierResponse",
        "description": "Response for outlier conversations."
      },
...
      "ThemeInsight": {
        "properties": {
          "theme": {
            "type": "string",
            "title": "Theme"
          },
          "cluster_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cluster Ids"
          },
          "frequency": {
            "type": "integer",
            "title": "Frequency"
          },
          "example_summaries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Example Summaries"
          }
        },
        "type": "object",
        "required": [
          "theme",
          "cluster_ids",
          "frequency",
          "example_summaries"
        ],
        "title": "ThemeInsight",
        "description": "Common theme across clusters."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}